// code-snippet-front/src/api-client/generated/models/CodeSnippetVO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CodeSnippetVO = {
    id?: string;
    title?: string;
    description?: string;
    content?: string;
    language?: string;
    copyCnt?: number;
    category?: string;
    updateTime?: string;
};